Initial Node setup
This commit adds linting, transpiling, bundling, and an editorconfig to
the project just to get us a basic lay of the land.

It also sets up the initial package.json with basic project information.

Finally, it creates an .npmignore file so that we can ignore the
compiled output from github while still including it in the npm
distribution.