Add support for modified proximity clauses
Some proximity clauses can be modified to adjust the distance that they
test against.  For example ADJ15 would look 15 tokens instead of the
default.

This adds support for these modifications.  Note that if there is
whitespace between the proximity operator and the number, that number is
actually interpreted as the term being compared.

Issue #3