Add support for fields and extensions
You can modify exactly what attribute of a document is being queried by
using extensions.  Some of these extensions will need to be made more
robust depending on the type of data they are expected to contain (e.g.
dates or numbers) but for now this adds basic extension support and logs
all types of extension that could be invoked.

Issue #3