Add support for wildcards
Wildcards (indicated by $) let the user specify an arbitry number of
generic characters after a word or quoted phrase.  Because wildcards
have an optional number that can appear after the $, we had to actually
incorporate that number as part of the tokenizing (rather than part of
the grammar).

The issue was that the way our grammar works, whitespace after clauses
is generally optional because we're relying on the tokenizer to split up
things like terms.  This means that by simply having a wildcard clause
with an optional `%number` our parser would not be able to tell the
difference between:

`word$ 5`

and

`word$5`

By making $5 and $ both singular "wildcard" tokens we avoid that kind of
ambiguity.

Issue #3