Add initial postprocessing
The nearley grammar we have defined so far does not produce a result
that is particularly usable.  It's helpful in the sense that it
validates a querystring, but not in the sense that it actually parses
that string into a useful structured object.

This commit begins the process of cleaning up the object that nearley
produces.  It "nulls" whitespace (which essentially removes it) and also
prevents the infinitely nested array structure that nearley generates by
default (by default, every rule creats a new nested array).

It turns out that nearley also creates arrays around parentheticals, so
this also removes a set of unnecessary parens around the conjunction
clause.

Issue #3