Clean root query postprocessing
Comments don't need to be in a special query field, but rather can just
be a typed clause.  This adjusts the structure of a query object to
always return a `content` array.

This also changes the way comments are postprocessed to make them a bit
cleaner / have a named type.