Restructure case insensitivity
We added support for case insensitivity in keywords, but actually went
a step too far and therefore implemented a feature against
specification.

The specification dictates that keywords can either be all capitalized
or all lowercase, but not a blend of the two.  This brings the case
insensitivity to that specification and removes support for detection of
mixed case keywords.

Issue #11